# src/main/java/fhict/boards/domain/dto/CommentRequest.java
package fhict.boards.domain.dto;

import jakarta.validation.constraints.Size;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@NoArgsConstructor
@AllArgsConstructor
public class CommentRequest {
    @Size(max = 500, message = "Comment must not exceed 500 characters.")
    private String body;
}